+++ NEWS.md
@@ -1,3 +1,36 @@
+# BuzzardsBay 0.1.0.9003
+2024-05-10
+
+* Update example dataset. In prior example data the Conductivity didn't match
+  in the DO and Cond. data and it should.
+* Update Details.txt processing to work with newer data that has different 
+  line endings.  It now works with any combination of CR and LF line endings.
+* Add test for consistent conductivity and dropped duplicate conductivity 
+  column.
+* Add test for consistent calibration start and end time in the two details
+  files and consolidated the start and end times in the metadata file.
+* Drop <data>_flag  columns.   "Flags" covers all flags.
+* Adopt standard "year-month-day" date, and "year-month-day h:m:s"
+  date time formats everywhere.
+* Add more checks for consistency against sites table and placements table.
+  * Logger SN's must be in placement table at the site indicated by the 
+    deployment path.
+  * Deployment date from path must match date of last observation in data.
+* Add range checks against sites table for DO, both Temperature Columns,
+  DO Pct. Saturation, and 
+* Add DO ratios to metadata - this is the 
+      "meter/titration value (mg/L)" divided by the "DO conc"
+       from the DO details.txt file created by HOBOware for both the 
+       start and end calibrations
+* Add `sites.csv` and `placements.csv` to example data and to files created
+  by `setup_example_data()` 
+* Cut off the pre and post calibration observations from the head and tail of 
+   the file. 
+* Add  "pct_calibrated", "n_records", "pct_immediate_rejection", and
+  "pct_flagged_for_review" to metadata.
+* Add warning if "pct_calibrated" is less than 95.
+
+
 # BuzzardsBay 0.1.0.9002
 
 * Add `README.Rmd` and `NEWS.md` files.